Remove all undesired usages of flat_map()
Even when assuming the inputs are always valid, we might parse things
wrongly.

Discarding those errors will not help us find the bug. For example:

```rust
    // prints [1, 2, 3], but we wanted [1, 2, 3, 4]
    // this is because we forgot to trim()
    //   :(
    println!(
        "{:?}",
        "1,2,3,4\n"
            .split(',')
            .flat_map(str::parse::<i32>)
            .collect::<Vec<_>>()
    );
```

We forgot to use trim() on the string, hence resulting in a bug.

These bugs are difficult to trace without the panics. flat_map() doesn't
panic so the errors are unnoticed.